+++ mix/backend.ex
@@ -1,0 +1,4 @@
+defmodule Currency.Cldr do
+  use Cldr, locales: ["en", "de"], default_locale: "en"
+
+end